+++ CHANGELOG.md
@@ -4,12 +4,24 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.9.6 - Unreleased
+## Version 0.9.7 - Unreleased
+
+### Changed
+* Changed image of Carl Sagan to one I know has a CC license. 
+* Added IPFS CIDs to all demo images.
+* Minor doc improvements.
+
+
+### Added
+* interpolate to Polygon and Color
+
+## Version 0.9.6 - Released 2022-08-10
 
 ### Added
 * The `bg_color` arg to `draw_header_text`
 * Implemented `kwimage.warp_image`
 * The `kwimage.Color.forimage` function is now public
+* Basic support for polygon interpolation.
 
 ### Fixed
 * Issue in Detections.draw where "cids" was not respected.